Prevent scrolling DateBar outside of image range
Most ways of scrolling already prevented this, but the mouse wheel was
not bounded in any way.

Note: you could still scroll for large amounts in a single step and
overcome the limitation, but even with holding the Shift modifier when
scrolling you can in practice only overshoot by a relatively small
amount.