Allow horizontal scrolling where appropriate.
As a follow up to d908ac10, I've reviewed scrollable areas in
KPhotoAlbum and determined that both the DateBarWidget and the DateEdit
lineedit in the annotation dialog should allow for horizontal scrolling.

As a sidenote: the DateEdit had an inverse scroll direction compared to
the corresponding date picker. This was rectified in this commit.