Fix image count in DateBar
When counting images, the earliest starting of two time ranges with the
same end time was skipped, leading to an image count that was too low.
This bug is easily demonstrated using the demo database, as seen in
the linked bug report.

Also, images with a start date that is exactly the same as the end of
the time range would be ignored. This bug may be triggered by an image
with a non-fuzzy date that was shot exactly on Dec. 31st at 23:59:59
when having the datebar units set to years.

CCBUG: 468045