Merge "Recreate Selected Thumbnails" and "Recalculate Checksums" functionality
If an image was changed visually, requiring the thumbnail to be refreshed, the
checksum should always be recreated as well.
If an image checksum was changed, there is always the potential for the
thumbnail to have become outdated.

Thus, there is no need to differentiate both functions. Checksum and
thumbnail should always be refreshed at the same time.