Break dependency of ImageDB on NewImageFinder
NewImageFinder has hard dependencies outside of DB, which won't work
when the database is moved into its own library.

Commit notes:

ImageDB::slotRescan():
The dirty() and totalChanged() signals are not
really needed because these are triggered anyway by the image insertions
to the database.

ImageDB::slotRecalcCheckSums():
As with slotRescan(), there's no need to trigger the signals from the
calling code. I kept the md5Map()->clear() even though it might not be
strictly needed - at least it avoids a load of lookups when the
information is rebuilt...

DB::ImageDB::imagesWithMD5Changed()
This is just a remnant of long deprecated funtctionality. I've just
removed it altogether.