Fix crash when AnnotationDialog is called from ViewerWidget
If the AnnotationDialog is called from the ViewerWidget (via the
"Annotate..." action in the context menu, ViewerWidget::editImage() is
called which calls MainWindow::configureImages().
Since MainWindow::configureImages() is a potentially long-running
function call, it is possible that the ViewerWidget was closed (and
deleted) before the function returns.
In this case, the execution returns to ViewerWidget::editImage of the
already deleted ViewerWidget object, leading to a crash.

By calling MainWindow::configureImages asynchronoulsy, the whole issue
is avoided.

BUG: 470889