Prevent inconsistent UI when thumbnail model is reset
If an action like "Collaps all stacks" is called, the thumbnail model is
reset, thus invalidating the selection.
Unfortunately, this does not trigger a selectionChanged signal because
the selection is reset and not cleared.

This would lead to an inconsistent UI, where actions that depend on an
existing selection would not be disabled, despite there not being a
selection.

BUG: 472109