Fix crash when the viewer gets a decoded pixmap while no image is set.
When viewing both images and videos, the situation might arise when the
ImageDisplay has not yet been set to an image after a transition from a
video to an image, yet the pixmapLoaded slot receives an image for the
previous video file.
For images, this is not an issue, because there's already a valid image
set (and it is checked whether the filename matches that of the received
pixmap).

BUG: 473587