Fix program freeze in VLC video backend
KPhotoAlbum would freeze up completely if the VLC video backend was used
and the viewer backend was closed while playback is active.
The reason for the freeze seems to be a deadlock inside libvlc if the
X11 window handle is already closed when playback is stopped (or if the
VLC player is released).

The fix is to release the VLC resources already when the CloseEvent is
issued, and the window still exists.