Discard images from annotation dialog if they are deleted.
This makes the annotation dialog behave like one would expect when an image
is deleted while being annotated:

Old behaviour: deleted images are still present in the annotation
dialog, leading to unclear semantics.

New behaviour: deleted images are removed from the dialog. The dialog is
closed when the last image is deleted.

Additionally, this fixes two crashes that are enabled by the old
behaviour:

Fix a crash when deleting the currently annotated image and then
toggling the fullscreen preview in the annotation window (#475387).

Fix a crash when deleting the currently annotated image and then
pressing delete in the annotation dialog (#475388).

BUG: 475387
BUG: 475388