Fix regression when deleting images
The introduction of AnnotationDialog::Dialog::slotDiscardImages()
reimplemented the things done in Dialog::slotDeleteImage(),
causing duplicated removal of images from the ImagePreviewWidget and
annotation Dialog. This caused a crash when deleting an image via the
delete button of the annotation dialog.

Fortunately, it turns out that the slotDiscardImages function does
everything slotDeleteImage() does, which means that the older code could
simply be removed.

Note: This patch removes the action "annotationdialog-delete-image",
which called Dialog::slotDeleteImage directly, and had a conflicting
default shortcut Ctrl+Delete (conflicting with Viewer -> delete image).
I believe the original intention for that action was to call
ImagePreviewWidget::slotDeleteImage(). Since the action was clearly not
in a working state, I thought it better to remove it entirely.

BUG: 475388