Fix crash when annotated image is deleted while full screen preview is active.
If the last image in the annotation dialog was deleted while the dialog
was showing the fullscreen preview, instead of closing the dialog as
expected, only the fullscreen preview was toggled.

This lead to the annotation dialog still showing the last image, even
though it should have been closed - thus triggering the same behaviour
as described in bug #475387.

BUG: 475387