Prevent image annotations in the ViewerWidget while in preview usage
It was up until now possible to change the rating of an image even if
the ViewerWidget was in preview mode (UsageType::CompactPreview or
UsageType::FullsizePreview).
This does not play well with the way changes are handled in the
annotation dialog, so I made the ViewerWidget defer all keypresses to
the parent widget while in preview mode.

This also means that the Qt::Key_Escape gets propagated and it's now
possible to close the fullscreen preview by pressing the escape key.