Fix compilation on Ubuntu 22.04 LTS
In commit b1f8247b8d99c29795340e9f83e4ae88456fdd13, the minimum ECM
version was set to 5.103, which was the version in Debian stable at that
time. Unfortunately, Ubuntu 22.04 LTS only ships ECM 5.92.

Fortunately, the reason for bumping the version in the first place was
KDEGitCommitHooks, which also works with version 5.92 (albeit with a
limited feature set).