Unify icon sizes in the browser category view.
So far, in TreeView with generic icons, folder icons would be much
larger than regular items, which seems kind of wonky to me.

To have the TreeView make some sense (compared to TreeView with custom
icons), I opted to keep the small icons in this case, so that for large
trees, one can have more entries compared to the TreeView with custom
icons (which should have reasonably big icons for them to be of use).