Fix crash when deleting a category while its category page is shown.
When the category page was shown while its category page is shown, one
could crash kphotoalbum by refreshing the view.

New behavior: Return to the home page when a deleted category page is
refreshed.

Note: even nicer would be to automatically go to the home page without a
manual refresh in this case.

BUG: 477532