Use the correct background color in viewer for zoomed images
When zooming in the viewer window, such that there is a border outside
of the image itself, that border was until now drawn black instead of
the QPalette::base() color.
The reason for this is that the zoomed image is actually a copy of the
loaded image with those borders.
Adding an alpha channel to the image fixes the issue.

It would possibly be a better fix to draw a scaled image in
ImageDisplay::paintEvent instead, but that would be a major rewrite of
the ImageDisplay code.

BUG: 478944