Start work on VideoThumbnailCache.
The goal of adding a VideoThumbnailCache class is to have a simple
API for video thumbnails, similar to the ThumbnailCache class.

Currently, most of the video thumbail code is in VideoThumbnails, which
is used both to load the video thumbnail files and to generate video
thumbnails for files that don't have them yet.

The VideoThumbnailCache is planned to have additional features compared
to the VideoThumbnails class:
 - invalidation of all thumbnails at once
 - invalidation of the thumbnails for one or more video files