Update Changelog
Add change log entry for the following change:
e32f2038 Don't allow non-existant directories with --db
b7367c78 Be more restricting in allowed filenames.
a38583d7 Remove unintended change.
b66ed4d4 Move and refactor per review comments.
0281fcc2 Bail with an error message on invalid input.
76590ed9 Handle directory passed on the command line.

Kudos to rsquared for doing most of the work on this!

Note: this is a port of the existing bugfix on the KF5/Qt5 branch
(master) to the kf6 branch.

BUG: 418647