Fix assertion when ImageScoutThread is used single-threaded
QMutexLocker(nullptr) obviously doesn't lock anything, but also does not
complain, as per its documented behaviour. Yet, calling
QMutexLocker::unlock() on such a mutex locker will trigger an assertion
because the mutex locker is not unlocked. Presumably, this assertion was
newly added in Qt6.

For that reason we now always initialize the ImageScoutThread with a
mutex, even if it is used with a single thread.
My reasoning is that the mutex will not significantly slow down anything
if only a single thread is accessing it.