WIP: remove UIDelegate dependency from SettingsData
This is an interim measure and adds a #warning to make sure we fix it
before release:
SettingsData::tTimeStamps() can be set to Always, Never, and Ask
SettingsData::trustTimestamps() is a wrapper around that and shows a
dialog is tTimeStamps() is set to Ask.

That dialog is removed by this commit.

For the next release, this code needs to be either reinstated (somewhere
else), or the value "Ask" has to be removed/replaced.