Fix crash when opening viewer after video playback failed
If KPhotoAlbum crashes during video playback, the next time when the
viewer widget is opened, an alternative video backend is selected.

Unfotunately, this caused a follow-up crash because of an
incorrect/unexpected type deduction where "auto" would be resolved to
QStringBuilder instead of QString, leading to a crash when the value was
used.