Escape category names for the "fast" db format as we did in KPA 5.x
Unfortunately, this broke during the Qt6/KF6 port. When reading a 5.x
database without this patch and the "fast" format is used, all tags from
all categories containing non-ASCII characters are ignored.

Apart from this, we have to reimplement this whole escaping, because in
the current form, it 1. only works for Latin-1 chars, not for the whole
Unicode range and 2. can produce invalid XML tag names, like ones
starting with a number (and they can't).