Don't draw histogram while dragging the DateBar
Most of the render time when redrawing the DateBarWidget is spent on
drawing the histogram. During dragging the DateBar, by omitting the
histogram, we can make the scrolling much smoother.

Currently, the fast scrolling feature can be toggled using the percent
key ('%'). This will be removed once the feature is merged.