Implement differenct histogram mode for DateBar
Instead of having differently sized bars, color the whole rectangle and
vary the color intensity.
Also, always write the number of images instead of only when the number
fits within the bar.

For testing, you can use the '$' key to toggle between histogram modes,
and use the '%' key to change the scaling (will be removed before the
next release).