Make sorting predictable in the merge duplicates dialog
The images were presented in DB order.  Now they are sorted by the pathname of the oldest copy of each image.

The order of the radio buttons for the copies is also now sorted by file creation time, so the oldest copy is always first and is selected by default.

A common use case: the user copies files from the camera into KPhotoAlbum but forgets to remove them from the camera, then later copies the same images (along with some new ones) into a new subfolder.  This can now be solved with 2-3 clicks assuming the older copies are kept.

This patch doesn't completely address the bug since there is still no quick way to select all of the newer copies in the example above.

CCBUG: 320480